requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.2=pypi_0
attrs=22.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blas=1.0=mkl
blinker=1.6.2=pypi_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py37haa95532_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
fonttools=4.38.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=pypi_0
imageio=2.27.0=pypi_0
importlib-metadata=6.3.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2023.0.0=h59b6b97_25923
jinja2=3.1.2=pypi_0
joblib=1.1.1=py37haa95532_0
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
mdurl=0.1.2=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37h196d8e1_0
mkl_fft=1.3.0=py37h46781fe_0
mkl_random=1.1.1=py37h47e9c7a_0
networkx=2.6.3=pypi_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.1=pypi_0
pandas=1.3.5=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=py37haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
protobuf=3.20.3=pypi_0
pyarrow=11.0.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.15.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.7.16=h6244533_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.3.0=pypi_0
requests=2.28.2=pypi_0
rich=13.3.4=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.6.2=py37h14eb087_0
setuptools=65.6.3=py37haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
streamlit=1.21.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=4.3=pypi_0
urllib3=1.26.15=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wheel=0.38.4=py37haa95532_0
wincertstore=0.2=py37haa95532_2
zipp=3.15.0=pypi_0
